pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
Pillow = "^9.1.0"
fonttools = "^4.33.3"
textwrap3 = "^0.9.2"
discord = "^1.7.3"
"discord-interactions.py" = "^0.0.8"
python-dotenv = "^0.20.0"
discord-ext-slash = "^0.9.2"
Flask = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"